seperating components into modules
