+++ tests/test_mutator.py
@@ -1,64 +1,13 @@
 # -*- coding: utf-8 -*-
+
 from __future__ import unicode_literals
+
 import sys
 
-from parso import parse
-
-from mutmut import mutate, count_mutations, ALL, Context, list_mutations, MutationID, array_subscript_pattern, function_call_pattern, ASTPattern
 import pytest
 
-
-@pytest.mark.skipif(sys.version_info < (3, 0), reason="Don't check Python 3 syntax in Python 2")
-def test_matches_py3():
-    node = parse('a: Optional[int] = 7\n').children[0].children[0].children[1].children[1].children[1].children[1]
-    assert not array_subscript_pattern.matches(node=node)
-
-
-def test_matches():
-    node = parse('from foo import bar').children[0]
-    assert not array_subscript_pattern.matches(node=node)
-    assert not function_call_pattern.matches(node=node)
-    assert not array_subscript_pattern.matches(node=node)
-    assert not function_call_pattern.matches(node=node)
-
-    node = parse('foo[bar]\n').children[0].children[0].children[1].children[1]
-    assert array_subscript_pattern.matches(node=node)
-
-    node = parse('foo(bar)\n').children[0].children[0].children[1].children[1]
-    assert function_call_pattern.matches(node=node)
-
-
-def test_ast_pattern_for_loop():
-    p = ASTPattern(
-        """
-for x in y:
-#   ^ n  ^ match
-    pass
-    # ^ x
-""",
-        x=dict(
-            of_type='simple_stmt',
-            marker_type='any',
-        ),
-        n=dict(
-            marker_type='name',
-        ),
-        match=dict(
-            marker_type='any',
-        )
-    )
-
-    n = parse("""for a in [1, 2, 3]:
-    if foo:
-        continue 
-""").children[0].children[3]
-    assert p.matches(node=n)
-
-    n = parse("""for a, b in [1, 2, 3]:
-    if foo:
-        continue 
-""").children[0].children[3]
-    assert p.matches(node=n)
+from mutmut.mutator import count_mutations, list_mutations, Context, mutate, \
+    ALL, MutationID
 
 
 @pytest.mark.parametrize(
